+++ nilearn/_assets/css/partials/navbar.css
