+++ nilearn/_assets/html/body_base.jinja
