+++ nilearn/_assets/html/glm/body_glm.jinja
