+++ nilearn/_assets/html/glm/partials/method_section.jinja
