+++ nilearn/_assets/html/glm/partials/navbar.jinja
