+++ nilearn/_assets/html/head.jinja
