+++ nilearn/_assets/html/masker/body_masker.jinja
