+++ nilearn/_assets/html/masker/body_nifti_labels_masker.jinja
