+++ nilearn/_assets/html/masker/body_nifti_maps_masker.jinja
