+++ nilearn/_assets/html/masker/body_nifti_spheres_masker.jinja
