+++ nilearn/_assets/html/masker/body_surface_maps_masker.jinja
