+++ nilearn/_assets/html/masker/body_surface_masker.jinja
