+++ nilearn/_assets/html/masker/partials/brainsprite.jinja
