+++ nilearn/_assets/html/masker/partials/brainsprite_opacity.jinja
