+++ nilearn/_assets/html/masker/partials/carousel.jinja
