+++ nilearn/_assets/html/masker/partials/figure.jinja
