+++ nilearn/_assets/html/plotting/connectome_plot.jinja
