+++ nilearn/_assets/html/plotting/surface_plot.jinja
