+++ nilearn/_assets/html/plotting/view_img.jinja
