+++ nilearn/_assets/js/carousel.js
