+++ nilearn/_assets/js/common-surface-plot-utils.js
