+++ nilearn/_assets/js/connectome-plot-utils.js
