+++ pyproject.toml
@@ -141,7 +141,7 @@ Homepage = "https://nilearn.github.io"
 [tool.codespell]
 builtin = "clear,rare,en-GB_to_en-US"
 ignore-words = ".github/codespell_ignore_words.txt"
-skip = "./.git,*.cff,*.bib,*/tmp,nilearn/assets/js/plotly-gl3d-latest.min.js,nilearn/assets/js/jquery.min.js,localizer_behavioural.tsv,.mypy_cache,env,venv,./doc/auto_examples,./doc/modules/generated/*,./doc/_build,./maint_tools/*.html,./examples/*/results,AUTHORS.rst"
+skip = "./.git,*.cff,*.bib,*/tmp,nilearn/_assets/js/plotly-gl3d-latest.min.js,nilearn/_assets/js/jquery.min.js,localizer_behavioural.tsv,.mypy_cache,env,venv,./doc/auto_examples,./doc/modules/generated/*,./doc/_build,./maint_tools/*.html,./examples/*/results,AUTHORS.rst"
 
 [tool.djlint]
 blank_line_after_tag = "load,extends,include,endblock,set"
@@ -294,7 +294,7 @@ allowed_importers = [
 ]
 name = "assets only imported by reporting, plotting, glm"
 protected_modules = [
-    "nilearn.assets"
+    "nilearn._assets"
 ]
 type = "protected"
